+++ src/style-spec/expression/definitions/interpolate.ts
@@ -135,6 +135,7 @@ class Interpolate implements Expression {
 
         if (outputType.kind !== 'number' &&
             outputType.kind !== 'color' &&
+            outputType.kind !== 'padding' &&
             !(
                 outputType.kind === 'array' &&
                 outputType.itemType.kind === 'number' &&