+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### ✨ Features and improvements
 
-- Add style swap's stylePatch support (#1238)
+- *...Add new stuff here...*
+- Update `icon-padding` symbol layout property to support asymmetric padding [#1289](https://github.com/maplibre/maplibre-gl-js/pull/1289)
 - Added `collaborativeGestures` option when instantiating map to prevent inadvertent scrolling/panning when navigating a page where map is embedded inline (#234)
+- Add style swap's stylePatch support (#1238)
 
 ### 🐞 Bug fixes
 
@@ -47,7 +49,7 @@
 
 - Fix missing `touchmove` in `MapTouchEvent["type"]` (#1131)
 - Type CustomLayerInterface renderingMode, onRemove, onAdd, and prerender optional (#1122)
- 
+
 ## 2.1.8-pre.3
 
 ### 🐞 Bug fixes